add module section to the project
