+++ index.html
@@ -4,8 +4,6 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-
-    <script src="index.js" defer></script>
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link
@@ -101,5 +99,7 @@
         <p>Copyright &copy;2022 Awesome Books</p>
       </div>
     </footer>
+
+    <script type="module" src="index.js"></script>
   </body>
 </html>
